Input: s6sa552 - add a driver for the Samsung A552 touchscreen controller
The S6SA552 touchscreen is a capacitive multi-touch controller for
mobile use. It connects via i2c at the address 0x48.

Introduce a basic driver, which can handle initialization, touch events
and power states.

At least the firmware for this IC on Galaxy S7 differs from S6SY761
in register layout and bits, as well as some missing registers/functions,
for example for retrieving the max X/Y coordinates and the amount
of TX/RX channels.

Signed-off-by: Ivaylo Ivanov <ivo.ivanov.ivanov1@gmail.com>